backend/requirements.txt
Flask~=3.0.3
Flask-Cors~=5.0.1
python-dotenv~=1.0.1
Werkzeug
gunicorn
pymongo~=4.10.1
google-generativeai
langchain-community~=0.3.14
langchain-core~=0.3.29
langchain-openai~=0.2.14
requests~=2.32.3
pdf2image~=1.17.0
protobuf~=5.29.4
pandas~=2.2.3
uvicorn~=0.34.2
fastapi~=0.115.12